requirements.txt
PySide6>=6.5.0
numpy>=1.24.0
scipy>=1.10.0
PyOpenGL>=3.1.7
PyOpenGL-accelerate>=3.1.7
pyqtgraph>=0.13.0
pyyaml>=6.0
trimesh>=3.21.0
pytest>=7.4.0
